Fix migrations

+++ database/migrations/2024_07_21_232618_remove_columns_from_analyses_table.php
@@ -14,7 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('analyses', function (Blueprint $table) {
-            $table->dropColumn('bofi_step_name');
+            if (Schema::hasColumn('analyses', 'bofi_step_name')) {
+                $table->dropColumn('bofi_step_name');
+            }
         });
     }
